Rename `types/contentlayer` to `types/docs`
